Update Camera.m
I found some issues while trying to update the 3D drawing of the Camera Object. Apparently this was caused by the deprecated variable 'h_camera3D' which seemed to be replaced by 'h_visualize' in the last versions of the toolbox. Now I can change the camera pose T and the drawing is updated correctly.

Thanks for the toolbox, it's amazing :)